Added support for mercurial
